Move many files to src folder to clean up repo.
